refactor: split the three remaining god services (pure moves, zero test edits)
The last structural item from the codebase grading: prospectService,
campaignService and luckyDrawService each sat near 1,200 lines. All
three now sit under 900, with the extracted seams chosen for cohesion,
not line count:

- prospectService 1237 → 883: the staff MUTATION surface
  (updateProspect — the H3/H4 hot zone — deleteProspect's
  transactional-outbox teardown, scheduleFollowUp, bulk delete) moves
  to prospectMutationService.js as a make*Ops factory, the same DI
  shape as the seven sibling factories the service already composes.
- campaignService 1165 → 898: the READ side (metrics, listing, detail,
  summary, analytics — everything that only SELECTs) moves to
  campaignReadService.js; campaignService re-exports the names so
  every controller/test import path is untouched (the P4-4 draw-guards
  precedent). The H1/H5 transactional write core stays put.
- luckyDrawService 1194 → 829: the boost-evidence collector + the
  prospect-facing draw-status projection move to
  luckyDrawStatusService.js, composed with the SAME entryEligibility
  predicate so freeze and status apply one eligibility rule.

Every function moved verbatim — no behaviour change. Proof: the full
unit tier (2237/2237) and the owning DB suites (campaign 133,
prospect/mutation 107, draw 146) pass with ZERO test edits; eslint
clean.
